+++ inst/create_sample_data.php
@@ -52,19 +52,19 @@ if ( $res===false ) die(__LINE__ . " - ".print_r($DBH->errorInfo(), true));
 // CATEGORIES_1
 //
 $query_categories_1_mysql   = 'INSERT INTO '.TBL_CATEGORIES_1.' '.
-                      '            (Category_1_id, Category_1_name, Category_1_description, Category_1_trashed) '.
-                      '     VALUES (0, "-- nessuna --", "", 0), '.
-                      '            (1, "SPES - Assistenza informatica", "Problematiche con PC, rete, sistema operativo, stampanti, configurazioni, telefonia, server, ecc...", 0), '.
-                      '            (2, "Sviluppo software", "Lavori di architettura e creazione codice", 0); ';
+                      '            (Category_1_name, Category_1_description, Category_1_trashed) '.
+                      '     VALUES ("-- nessuna --", "", 0), '.
+                      '            ("SPES - Assistenza informatica", "Problematiche con PC, rete, sistema operativo, stampanti, configurazioni, telefonia, server, ecc...", 0), '.
+                      '            ("Sviluppo software", "Lavori di architettura e creazione codice", 0); ';
 $query_categories_1_sqlite  = 'INSERT INTO '.TBL_CATEGORIES_1.' '.
-                      '            (Category_1_id, Category_1_name, Category_1_description, Category_1_trashed) '.
-                      '     VALUES (0, "-- nessuna --", "", 0); '.
+                      '            (Category_1_name, Category_1_description, Category_1_trashed) '.
+                      '     VALUES ("-- nessuna --", "", 0); '.
                       '        INSERT INTO '.TBL_CATEGORIES_1.' '.
-                      '            (Category_1_id, Category_1_name, Category_1_description, Category_1_trashed) '.
-                      '     VALUES (2, "Sviluppo software", "Lavori di architettura e creazione codice", 0); '.
+                      '            (Category_1_name, Category_1_description, Category_1_trashed) '.
+                      '     VALUES ("Sviluppo software", "Lavori di architettura e creazione codice", 0); '.
                       '        INSERT INTO '.TBL_CATEGORIES_1.' '.
-                      '            (Category_1_id, Category_1_name, Category_1_description, Category_1_trashed) '.
-                      '     VALUES (1, "SPES - Assistenza informatica", "Problematiche con PC, rete, sistema operativo, stampanti, configurazioni, telefonia, server, ecc...", 0);';
+                      '            (Category_1_name, Category_1_description, Category_1_trashed) '.
+                      '     VALUES ("SPES - Assistenza informatica", "Problematiche con PC, rete, sistema operativo, stampanti, configurazioni, telefonia, server, ecc...", 0);';
 if ( $config['DB']['type']==='mysql' ) $query_categories_1 = $query_categories_1_mysql;
 else                                   $query_categories_1 = $query_categories_1_sqlite;
 $res = $DBH->exec($query_categories_1);                                      // insert Categories_A data
@@ -75,18 +75,27 @@ if ( $res===false ) die(__LINE__ . " - ".print_r($DBH->errorInfo(), true));
 // CATEGORIES_2
 //
 $query_categories_2_mysql  = 'INSERT INTO '.TBL_CATEGORIES_2.' '.
-                      '            (Category_2_id, Category_2_id_Category_1, Category_2_name, Category_2_description, Category_2_trashed) '.
-                      '     VALUES (1, 1, "Disco o memoria danneggiata", "Problematiche di varia natura su disco fisso o altro sistema di memorizzazione", 0), '.
-                      '            (2, 1, "Configurazione/verifica gateway/DNS/NATP", "Configurazione di apparati di rete per accessi e controlli", 0), '.
-                      '            (3, 2, "Sviluppo GUI", "Architettura e sviluppo interfaccia utente", 0), '.
-                      '            (4, 2, "Sviluppo logiche server-side", "Architettura e sviluppo di software per interazione remota con il server", 0), '.
-                      '            (0, 0, "-- nessuna --", "", 0); ';
+                      '            (Category_2_id_Category_1, Category_2_name, Category_2_description, Category_2_trashed) '.
+                      '     VALUES (1, "-- nessuna --", "", 0), '.
+                      '            (2, "Disco o memoria danneggiata", "Problematiche di varia natura su disco fisso o altro sistema di memorizzazione", 0), '.
+                      '            (2, "Configurazione/verifica gateway/DNS/NATP", "Configurazione di apparati di rete per accessi e controlli", 0), '.
+                      '            (3, "Sviluppo GUI", "Architettura e sviluppo interfaccia utente", 0), '.
+                      '            (3, "Sviluppo logiche server-side", "Architettura e sviluppo di software per interazione remota con il server", 0);';
 $query_categories_2_sqlite = 'INSERT INTO '.TBL_CATEGORIES_2.' '.
-                      '            SELECT 1 AS "Category_2_id", 1 AS "Category_2_id_Category_1", "Disco o memoria danneggiata" AS "Category_2_name", "Problematiche di varia natura su disco fisso o altro sistema di memorizzazione" AS "Category_2_description", 0 AS "Category_2_trashed" '.
-                      '      UNION SELECT 2, 1, "Configurazione/verifica gateway/DNS/NATP", "Configurazione di apparati di rete per accessi e controlli", 0 '.
-                      '      UNION SELECT 3, 2, "Sviluppo GUI", "Architettura e sviluppo interfaccia utente", 0 '.
-                      '      UNION SELECT 4, 2, "Sviluppo logiche server-side", "Architettura e sviluppo di software per interazione remota con il server", 0 '.
-                      '      UNION SELECT 0, 0, "-- nessuna --", "", 0 ';
+                      '                   (Category_2_id_Category_1, Category_2_name, Category_2_description, Category_2_trashed) '.
+                      '            VALUES (1, "-- nessuna --", "", 0 ); '.
+                      '       INSERT INTO '.TBL_CATEGORIES_2.' '.
+                      '                   (Category_2_id_Category_1, Category_2_name, Category_2_description, Category_2_trashed) '.
+                      '            VALUES (2, "Configurazione/verifica gateway/DNS/NATP", "Configurazione di apparati di rete per accessi e controlli", 0); '.
+                      '       INSERT INTO '.TBL_CATEGORIES_2.' '.
+                      '                   (Category_2_id_Category_1, Category_2_name, Category_2_description, Category_2_trashed) '.
+                      '            VALUES (2, "Sviluppo GUI", "Architettura e sviluppo interfaccia utente", 0); ' .
+                      '       INSERT INTO '.TBL_CATEGORIES_2.' '.
+                      '                   (Category_2_id_Category_1, Category_2_name, Category_2_description, Category_2_trashed) '.
+                      '            VALUES (2, "Sviluppo logiche server-side", "Architettura e sviluppo di software per interazione remota con il server", 0); '.
+                      '       INSERT INTO '.TBL_CATEGORIES_2.' '.
+                      '                   (Category_2_id_Category_1, Category_2_name, Category_2_description, Category_2_trashed) '.
+                      '            VALUES (2, "Disco o memoria danneggiata", "Problematiche di varia natura su disco fisso o altro sistema di memorizzazione", 0); ';
 if ( $config['DB']['type']==='mysql' ) $query_categories_2 = $query_categories_2_mysql;
 else                                   $query_categories_2 = $query_categories_2_sqlite;
 $res= $DBH->exec($query_categories_2);                                      // insert Categories_B data