+++ lib/requests_categories.php
@@ -19,10 +19,18 @@ function dispatch_request($request)
 
             $CAT->load();                               // get categories according to parameters preiousely set
 
+            $tbl_name = 'Category_' . $CAT->level;
+            foreach ( $CAT->category_data as $key => $value ) {
+                $retval['data'][] = array( 'id'          => $value[ $tbl_name . '_id'],
+                                           'name'        => $value[ $tbl_name . '_name'],
+                                           'description' => $value[ $tbl_name . '_description'] );
+            }
+
             // generate a new nonce/timestamp for ajax call
             $json_nonce = generate_json_values($command, 0, session_id(), $config['SALT'], $config['HASH_ALG']);
             $retval['new_timestamp'] = $json_nonce['timestamp'];
             $retval['new_nonce']     = $json_nonce['nonce'];
+
             return;
 
             break;